New ERT Shuttle for quick deployment of ERT forces
Meant to be used for when ERT is needed on station ASA-NOW

Features;
Quick speed
6 Jump seat for ERT forces to occupy while they ride to the station
No atmos on the shuttle, internals will be needed.

The concept is simple, get the ERT squad on board and FTL to station and have the pilot get close then the ERT squad can jump from the run boards and do a quick EVA to station before going about their objective.