Corpsman Addition
Whew. That was a lot of work.
This finally makes a role separation between corpsman and brigmedic, allowing the brigmedic to stay in brig and the corpsman to run out and perform the other half of the brigmedic's duties.

Basically, brig is doctor/chemist, stuck in brig
Corps is para/on field triage meant to retrieve & heal seccies in combat conditions