Passive equipment implementation + active equipment fixes
Night vision is really the only currently used passive and it was exceedingly straightforward converting it to a component registry system. This may have issues regarding components being added or removed when already existing and so a delta might be appropriate to add as well. E.g. movement modifier systems such as a leg boost or thrusters.