Add Library access, doors, crates, and locker
This is in preparation for further library additions, but should stand on its
own as a reviewable unit. Further responsibilities for librarians as part of
the library will likely need some level of security, so this changeset brings
them the standard set of available access control objects.

The general door resprite push should go through next month; until then,
the hydroponics door sprites are very obviously service but aren't set
to have specific branding for hydroponics.