add(paperwork): Add [datetime] tag to easily add date and time.
This adds a new paperwork markup tag [datetime] to make adding the date
and time a single click, just like check boxes or other items. It should
hopefully cut down a bit of process for paperwork enjoyers since opening
the PDA and either writing the date and time down, or knowing that you
can copy that somehow feels a bit cumbersome and I want paperwork
non-enjoyers to engage with someone doing paperwork on an more equal
level.

It also adjusts the wordage across all items to "Date and Time", even
though the PDA (and this PR) uses time and date as I didn't want to
modify a ton of upstream code. I'm open to reworking this to fit ISO
8601 or RP equivalent, though this is just the output on the paperwork.
Depending on the goal here I can look at possibly changing it upstream
unless you want to add that extra work.

Another item is that there are two functions used to get the time and/or
date, one of those returns both, so I've removed any usage of the
singular return function and the function itself as they both do
seemingly slightly different ways of handling Starlight's time. I'm open
to revert this part or reworking another way.